➕🎨  Bundle Remote via Webpack
