+++ src/main/integrations/companion-server/remote/src/css/_material-tabs.scss
@@ -1,0 +1,85 @@
+$number-of-tabs: 3;
+
+.material-header {
+  position: relative;
+  background: rgba(20, 21, 22, 80%);
+  backdrop-filter: blur(20px);
+
+  .material-tabs {
+    position: relative;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -ms-flex-pack: distribute;
+    justify-content: space-around;
+
+    padding: 0;
+    border-bottom: 1px solid #7c7c7c;
+
+    & > a {
+      position: relative;
+      display: inline-block;
+      text-decoration: none;
+      padding: 16px;
+      text-transform: uppercase;
+      font-size: 14px;
+      font-weight: 600;
+      color: #cccccc;
+      text-align: center;
+
+      width: 120px;
+
+      .active {
+        font-weight: 700;
+        outline: none;
+      }
+    }
+
+    & > a:not(.active):hover {
+      background-color: inherit;
+      color: #7c848a;
+    }
+
+    @for $i from 1 through $number-of-tabs {
+      .material-tab:nth-of-type(#{$i}).active ~ span.active-line {
+        left: calc((100% / $number-of-tabs) * #{$i - 1});
+        width: calc(100% / $number-of-tabs);
+      }
+    }
+  }
+
+  .active-line {
+    position: absolute;
+    z-index: 10;
+    bottom: 0;
+    height: 3px;
+    background: #f5f5f5;
+    display: block;
+    left: 0;
+    transition: left 0.2s ease;
+    -webkit-transition: left 0.2s ease;
+  }
+
+  /* blurred background */
+  .tab-contents {
+    position: relative;
+  }
+
+  .tab-contents::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgb(35, 36, 37, 0.8);
+    backdrop-filter: blur(20px);
+
+    z-index: -1;
+  }
+
+  .material-tab-content {
+    height: calc(100vh - var(--bottom-draw-height) * 2);
+    padding: 16px;
+  }
+}
