+++ src/main/integrations/companion-server/remote/src/css/_queue.scss
@@ -1,0 +1,109 @@
+#queue {
+  display: flex;
+  flex-direction: column;
+  overflow-y: auto;
+
+
+  .queue-separator {
+    padding: 20px 0;
+    border-bottom: 1px solid grey;
+  }
+
+  .queue-item {
+    display: flex;
+    flex-direction: row;
+    justify-content: start;
+    align-items: center;
+    gap: 16px;
+    padding-right: 4px;
+    cursor: pointer;
+
+    border-bottom: 1px solid #464646;
+
+    &.selected {
+      background: rgba(255, 255, 255, 0.2);
+    }
+
+    &:hover {
+      background-color: rgba(255, 255, 255, 0.1);
+    }
+
+
+    .queue-item-album {
+      height: 64px;
+      position: relative;
+
+      /* Display a icon on the selected item */
+      /* .selected &::before {
+        content: "";
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background: rgba(0, 0, 0, 0.2);
+        z-index: 1;
+      }
+
+      .selected &::after {
+        content: "";
+        position: absolute;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        background: url(#volume) no-repeat center;
+        background-size: 50%;
+        z-index: 2;
+      } */
+    }
+
+    .queue-albumart {
+      aspect-ratio: 1 / 1 !important;
+      max-height: 100%;
+      max-width: 64px;
+      object-fit: cover;
+    }
+
+    .queue-item-details {
+      display: flex;
+      flex-direction: column;
+      max-width: calc(100% - 64px - 32px - 40px);
+
+      .queue-item-title {
+        font-size: 16px;
+        font-weight: 500;
+        line-height: 1.2;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        overflow: hidden;
+      }
+
+      .queue-item-artist {
+        font-size: 12px;
+        font-weight: 400;
+        line-height: 1;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        overflow: hidden;
+
+        color:rgba(255, 255, 255, 0.7);
+      }
+    }
+
+    .queue-item-end {
+      margin-left: auto;
+      max-width: 40px;
+
+      .queue-item-duration {
+        font-size: 12px;
+        font-weight: 400;
+        line-height: 1;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        overflow: hidden;
+        color:rgba(255, 255, 255, 0.7);
+      }
+    }
+  }
+}
