+++ webpack.rules.ts
@@ -21,5 +21,16 @@ export const rules: Required<ModuleOptions>["rules"] = [
   {
     test: /\.script\.js$/,
     type: "asset/source"
+  },
+  {
+    test: /\.s[ac]ss$/i,
+    use: [
+      // Creates `style` nodes from JS strings
+      "style-loader",
+      // Translates CSS into CommonJS
+      "css-loader",
+      // Compiles Sass to CSS
+      "sass-loader"
+    ]
   }
 ];
